fix(server): verify vendored runner deps instead of bundling them
Greptile flagged a real regression in the esbuild-bundling approach from
the previous commit: several paperclip-runner modules resolve sibling
build artifacts (the native runnerd binary, ACPX/OpenCode CJS sidecar
scripts under dist/cli, JSON replay fixtures under a top-level protocol/
directory) via import.meta.url-relative filesystem paths rather than JS
imports, at whatever nesting depth their source file happens to sit at.
Bundling the runner's entry points collapses and rearranges that layout,
silently breaking those lookups -- exactly the "green CI, broken
production" failure shape this fix exists to prevent, just relocated.

Revert the vendor step to the original, proven `cp -R` of the runner's
whole compiled dist/ tree -- it preserves every relative path exactly,
so none of those lookups break. Keep the goal from the previous commit
(derive the required dependency set instead of hand-copying it) as a
separate, purely additive check: scripts/verify-runner-vendor-dependencies.mjs
runs esbuild with `write: false` against the same two entry points
purely as a module-graph analysis tool -- nothing is written to disk --
and fails the build with a precise, actionable error if any package the
bundle would actually need isn't declared in server/package.json.

This is also more precise than a blanket "mirror every dependency"
rule: paperclip-runner declares dependencies (react-markdown, the
codex/opencode CLI packages, ...) that only its unrelated ./react and
./browser export subpaths use. Running the real check against this
repo's manifests today confirms it: those show up as extraneous and are
correctly left unflagged, while a synthetic "smol-toml missing" fixture
correctly throws.
